More tests, add a FIXME re: HTTP response codes
Currently, most HTTP response codes are incorrectly treated as though
they were integers, summing the codes in each bin. So if there are three
200 OK response codes in a bin, the "total" will be 600, which is pretty
meaningless. It would be better to count up the total number of each
response code in each bin, such that:

    >>> bin.stats['HTTP response code']
    {
      '200': 3,
      '404': 0,
      '503': 0,
    }

or something to that effect. Alternatively, since these numeric response
codes are unique among the list of typical Grinder statistics, it could
also work to log them as a top-level statistic, i.e.:

    >>> bin.stats['200']
    3
    >>> bin.stats['404']
    0
    >>> bin.stats['503']
    0

Something to consider for a future version, anyway...